@@ -1,13 +1,14 @@
 {
   "name": "create-spl-app",
-  "version": "0.0.3",
+  "version": "0.0.4",
   "keywords": [
     "spl"
   ],
+  "type": "module",
   "description": "Create SPLs with no build configuration.",
   "license": "MIT",
   "engines": {
-    "node": ">=14"
+    "node": ">=20"
   },
   "bin": {
     "create-spl-app": "./index.js"
@@ -16,21 +17,11 @@
     "test": "cross-env FORCE_COLOR=true jest"
   },
   "dependencies": {
-    "chalk": "^4.1.2",
-    "commander": "^4.1.1",
     "cross-spawn": "^7.0.3",
-    "envinfo": "^7.8.1",
-    "fs-extra": "^10.0.0",
-    "hyperquest": "^2.1.3",
+    "kolorist": "^1.8.0",
+    "minimist": "^1.2.8",
     "prompts": "^2.4.2",
-    "semver": "^7.3.5",
-    "tar-pack": "^3.4.1",
-    "tmp": "^0.2.1",
-    "validate-npm-package-name": "^3.0.0"
-  },
-  "devDependencies": {
-    "cross-env": "^7.0.3",
-    "jest": "^27.4.3"
+    "semver": "^7.3.5"
   },
   "author": {
     "name": "Victor Lamas",
