@@ -40,16 +40,15 @@ const currentNodeVersion = process.versions.node;
 const semver = currentNodeVersion.split('.');
 const major = semver[0];
 
-if (major < 14) {
+if (major < 20) {
   console.error(
     'You are running Node ' +
-      currentNodeVersion +
-      '.\n' +
-      'Create React App requires Node 14 or higher. \n' +
-      'Please update your version of Node.'
+    currentNodeVersion +
+    '.\n' +
+    'Create SPL App requires Node 20 or higher. \n' +
+    'Please update your version of Node.'
   );
   process.exit(1);
 }
 
-const { init } = require('./createSplApp');
-init();
+import "./createSplApp.js";
